+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 August 2023
 
+- Add toggle option for toolbar visibility.
 - Add basic support for adding equipment and personnel to units.
 
 July 2023